==> 01_B_process_trial_info.ipynb <==
import pandas as pd
import pickle
import numpy as np
import xml.etree.cElementTree as ET
import os
import re
from tqdm import tqdm
import csv
import requests

payment = pickle.load(open('./data/payment_dict.pkl', 'rb'))
trial_list = list(payment.keys())

def get_icd_from_nih(disease_name):
	prefix = 'https://clinicaltables.nlm.nih.gov/api/icd10cm/v3/search?sf=code,name&terms='
	url = prefix + disease_name 
	response = requests.get(url)
	text = response.text 
	if text == '[0,[],null,[]]':
		return [] 
	text = text[1:-1]
	idx1 = text.find('[')
	idx2 = text.find(']')
	codes = text[idx1+1:idx2].split(',')
	codes = [i[1:-1] for i in codes]
	return codes

def process_trial(path):
    cur_trial = ET.ElementTree(file=path)
    cur_dict = {}
    if cur_trial.find('start_date') == None:
        return None
    else:
        cur_str = cur_trial.find('start_date').text
        # Use regex to find year
        cur_dict['start_date'] = pd.to_datetime(cur_str).strftime('%Y-%m-%d')
        
    if cur_trial.find('completion_date') == None:
        cur_dict['completion_date'] = np.nan
    else:
        cur_str = cur_trial.find('completion_date').text
        cur_dict['completion_date'] = pd.to_datetime(cur_str).strftime('%Y-%m-%d')
    
    # phase
    if cur_trial.find('phase') == None:
        cur_dict['phase'] = np.nan
    else:
        cur_dict['phase'] = cur_trial.find('phase').text
    
    cur_cond = []
    for x in cur_trial.findall('condition_browse/mesh_term'):
        cur_cond += get_icd_from_nih(x.text)
        
    cur_dict['condition'] = list(set(cur_cond))
    
    sum_text = cur_trial.find('brief_summary/textblock').text
    # Remove all \r\n, \n and extra spaces
    sum_text = re.sub(r'[\r\n]+', ' ', sum_text)
    sum_text = re.sub(r' +', ' ', sum_text)
    cur_dict['summary'] = sum_text
    
    if cur_trial.find('eligibility/criteria/textblock') == None:
        return None
    else:
        cur_dict['criteria'] = cur_trial.find('eligibility/criteria/textblock').text
    return cur_dict    

trial_dict = {}
for path, subdirs, files in tqdm(os.walk('./data/AllPublicXML/')):
    for name in files:
        if name[:-4] in trial_list:
            cur_trial = os.path.join(path, name)
            cur_res = process_trial(cur_trial)
            if cur_res != None:
                res = process_trial(cur_trial)
                if res != None:
                    trial_dict[name[:-4]] = process_trial(cur_trial)
                else:
                    continue
            else:
                continue
        else:
            continue


trial_dict

trial_dict['NCT00089245']

pickle.dump(trial_dict, open('./data/trial_info.pkl', 'wb'))

# Generate parser input csv
rows = []
for each_trial in trial_dict:
    rows.append([each_trial, '', '', trial_dict[each_trial]['condition'], trial_dict[each_trial]['criteria']])
with open('./data/trial_info.csv', 'w', newline='') as csvfile:
    writer = csv.writer(csvfile)
    writer.writerow(['#nct_id', 'title', 'has_us_facility', 'conditions', 'eligibility_criteria'])
    writer.writerows(rows)

trial_dict['NCT00089245']
