==> 01_C_process_trial_criteria_embd.ipynb <==
import numpy as np
import pandas as pd
import openai
import numpy as np
from numpy import dot
from numpy.linalg import norm
from tqdm import tqdm
from transformers import pipeline, AutoTokenizer, AutoModelForCausalLM, AutoModel
import pickle


with open('./data/ie_extracted_clinical_trials.tsv','rb') as f:
    df = pd.read_csv(f,sep='\t')
    df = df[df['eligibility_type'] == 'inclusion']

df['criterion'].apply(lambda x: len(x.split())).sum()

model = AutoModel.from_pretrained("medicalai/ClinicalBERT", trust_remote_code=True)
tokenizer = AutoTokenizer.from_pretrained("medicalai/ClinicalBERT", trust_remote_code=True, model_max_length=512, truncation=True, max_length=512)
extractor = pipeline(model=model, tokenizer=tokenizer, task="feature-extraction")

trial2criterion = {}
trial2embd = {}
for idx, row in tqdm(df.iterrows()):
    cur_nct = row['#nct_id']
    if cur_nct not in trial2criterion:
        trial2criterion[cur_nct] = []
        trial2embd[cur_nct] = []
    trial2criterion[cur_nct].append(row['criterion'])
    
    # Extract embeddings from text row['criterion']
    cur_embd = extractor(row['criterion'], truncation=True, max_length=512)
    
    # Generate the sentence embedding
    cur_embd = np.array(cur_embd).squeeze().mean(0)
    trial2embd[cur_nct].append(cur_embd)
    

pickle.dump(trial2criterion, open('./data/trial2criterion.pkl','wb'))
pickle.dump(trial2embd, open('./data/trial2embd.pkl','wb'))

trial2pool = {}
for trial in tqdm(trial2embd):
    # Get the average embedding for each trial
    trial2pool[trial] = np.array(trial2embd[trial]).mean(0)

pickle.dump(trial2pool, open('./data/trial2pool.pkl','wb'))
