==> 01_D_process_trial_summary_embd.ipynb <==
import numpy as np
import pandas as pd
import numpy as np
from numpy import dot
from numpy.linalg import norm
from tqdm import tqdm
from transformers import pipeline, AutoTokenizer, AutoModelForCausalLM, AutoModel
import pickle
import pyotp

trial_dict = pickle.load(open('./data/trial_info.pkl', 'rb'))
icd_list = pickle.load(open('./data/icd_list.pkl', 'rb'))

trial_dict['NCT00004451']

trial_condition = {}
for each in trial_dict:
    cur_cond = np.zeros(len(icd_list))
    for each_cond in trial_dict[each]['condition']:
        cur_name = each_cond[:3]
        if cur_name not in icd_list:
            continue
        cur_cond[icd_list.index(cur_name)] = 1
    trial_condition[each] = cur_cond

trial_summary = {}
model = AutoModel.from_pretrained("medicalai/ClinicalBERT", trust_remote_code=True)
tokenizer = AutoTokenizer.from_pretrained("medicalai/ClinicalBERT", trust_remote_code=True, model_max_length=512, truncation=True, max_length=512)
extractor = pipeline(model=model, tokenizer=tokenizer, task="feature-extraction")
for each in trial_dict:
    cur_sum = trial_dict[each]['summary']
    cur_embd = extractor(cur_sum, truncation=True, max_length=512)
    
    # Generate the sentence embedding
    cur_embd = np.array(cur_embd).squeeze().mean(0)
    trial_summary[each] = cur_embd

pickle.dump(trial_condition, open('./data/trial_condition.pkl', 'wb'))
pickle.dump(trial_summary, open('./data/trial_summary.pkl', 'wb'))
