==> 01_E_process_claims_data.ipynb <==
import pandas as pd
import pickle
import numpy as np

icd_df = pd.read_csv('./data/filtered_icd.csv')

unique_icd = list(icd_df['icd'].unique())
unique_icd.sort()
print(len(unique_icd))
pickle.dump(unique_icd, open('./data/filtered_icd.pkl', 'wb'))

# Convert each year's data to dict {npi: vector}
unique_icd = list(icd_df['icd'].unique())
unique_icd.sort()
for year in range(2016, 2021):
    cur_df = icd_df[icd_df['year'] == year]
    cur_df = cur_df.pivot(index='npi', columns='icd', values='cnt')
    # Add the missing columns
    missing_icd = list(set(unique_icd) - set(cur_df.columns))
    cur_df = cur_df.reindex(columns=sorted(cur_df.columns.tolist() + missing_icd))
    cur_df = cur_df.fillna(0)
    
    # Convert to dict of arrays
    cur_dict = {}
    for npi in cur_df.index:
        cur_dict[npi] = np.array(cur_df.loc[npi])
    pickle.dump(cur_dict, open('../data/icd_dict_{}.pkl'.format(year), 'wb'))

icd_list = list(cur_df.columns)
pickle.dump(icd_list, open('../data/icd_list.pkl', 'wb'))

cpt_df = pd.read_csv('./data/filtered_cpt.csv')
unique_cpt = list(cpt_df['cpt'].unique())
unique_cpt.sort()
for year in range(2016, 2021):
    cur_df = cpt_df[cpt_df['year'] == year]
    cur_df = cur_df.pivot(index='npi', columns='cpt', values='cnt')
    missing_cpt = list(set(unique_cpt) - set(cur_df.columns))
    cur_df = cur_df.reindex(columns=sorted(cur_df.columns.tolist() + missing_cpt))
    cur_df = cur_df.fillna(0)
    cur_dict = {}
    for npi in cur_df.index:
        cur_dict[npi] = np.array(cur_df.loc[npi])
    pickle.dump(cur_dict, open('./data/cpt_dict_{}.pkl'.format(year), 'wb'))
