==> 02_A_gen_trial_npi_relation.ipynb <==
import pandas as pd
import numpy as np
import pickle
from tqdm import tqdm

payment = pickle.load(open('./data/payment_dict.pkl', 'rb'))
trial_info = pickle.load(open('./data/trial_info.pkl', 'rb'))
trial2pool = pickle.load(open('./data/trial2pool.pkl', 'rb'))
npi_codes = pickle.load(open('./data/npi_codes.pkl', 'rb'))

trial_list = list(trial2pool.keys())

# Sort the trial list by the trial start date
trial_list.sort(key=lambda x: trial_info[x]['start_date'])
trial2idx = {trial_list[i]: i for i in range(len(trial_list))}

npi_list = set()
npi_code_set = set([each_key for each_key in npi_codes.keys()])
trial2npi = {}
for each_trial in tqdm(trial_list):
    cur_list = set(payment[each_trial][0].keys()).intersection(npi_code_set)
    for each_hos in payment[each_trial][1]:
        cur_list.update(set(payment[each_trial][1][each_hos].keys()).intersection(npi_code_set))
    for each_ins in payment[each_trial][2]:
        cur_list.update(set(payment[each_trial][2][each_ins].keys()).intersection(npi_code_set))
    trial2npi[each_trial] = list(cur_list)
    npi_list.update(cur_list)
npi_list = list(npi_list)
npi_list.sort()
npi2idx = {npi_list[i]: i for i in range(len(npi_list))}


npi2trial = {}
for each_trial in trial2npi:
    for each_npi in trial2npi[each_trial]:
        if each_npi not in npi2trial:
            npi2trial[each_npi] = []
        npi2trial[each_npi].append(each_trial)


pickle.dump(trial2idx, open('./data/trial2idx.pkl', 'wb'))
pickle.dump(npi2idx, open('./data/npi2idx.pkl', 'wb'))
pickle.dump(trial2npi, open('./data/trial2npi.pkl', 'wb'))
pickle.dump(trial_list, open('./data/trial_list.pkl', 'wb'))
pickle.dump(npi2trial, open('./data/npi2trial.pkl', 'wb'))

npi2trial['1114973203']
