==> 05_A_get_competing_trial.ipynb <==
import pickle
import numpy as np
from datetime import datetime

trial_info = pickle.load(open('./data/trial_info.pkl', 'rb'))
npi2trial = pickle.load(open('./data/npi2trial.pkl', 'rb'))
trial2npi = pickle.load(open('./data/trial2npi.pkl', 'rb'))
test_list = pickle.load(open('./data/zero_shot/test_list.pkl', 'rb'))

competing_dict = {}
cnt = 0
start_date = []
for each_trial in test_list:
    cur_start_date = trial_info[each_trial]['start_date']
    cur_start_date = datetime.strptime(cur_start_date, '%Y-%m-%d')
    competing_dict[each_trial] = {}
    for each_npi in npi2trial:
        competing_dict[each_trial][each_npi] = 0
    for each_npi in trial2npi[each_trial]:
        for each_trial2 in npi2trial[each_npi]:
            compete_trial_end_date = trial_info[each_trial2]['completion_date']
            if str(compete_trial_end_date) == 'nan':
                continue
            compete_trial_end_date = datetime.strptime(compete_trial_end_date, '%Y-%m-%d')
            if cur_start_date < compete_trial_end_date:
                competing_dict[each_trial][each_npi] += 1
                cnt += 1
    start_date.append(cur_start_date)

pickle.dump(competing_dict, open('./data/zero_shot/competing_dict.pkl', 'wb'))
